pyproject.toml
[project]
name = "pm-attendance-tracker"
version = "0.1.1"
description = "discord bot for tracking meeting attendance"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord-py>=2.7.1",
]

[project.scripts]
pm-attendance-tracker = "pm_attendance_tracker:main"

[build-system]
requires = ["uv_build>=0.11.6,<0.12.0"]
build-backend = "uv_build"


[dependency-groups]
dev = [
    "ruff>=0.15.10",
]
